feat: preload visitor stats and smooth initial counter animation
