+++ src/components/StatsWidget/index.tsx
@@ -1,7 +1,8 @@
 import { getAllPosts } from '@/utils/post-util';
+import { getVisitorStats } from '@/utils/stats-util';
 import { StatsWidgetClient } from './Client';
 
 export const StatsWidget = async () => {
-  const posts = await getAllPosts();
-  return <StatsWidgetClient postCount={posts.length} />;
+  const [posts, initialStats] = await Promise.all([getAllPosts(), getVisitorStats()]);
+  return <StatsWidgetClient initialStats={initialStats} postCount={posts.length} />;
 };
